refactor: move name input group error hook into its own file
